% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validators.R
\name{tif_tokens_validate}
\alias{tif_tokens_validate}
\title{Validate Tokens Object}
\usage{
tif_tokens_validate(tokens, warn = FALSE)
}
\arguments{
\item{tokens}{a tokens object to test for validity}

\item{warn}{logical. Should the function produce a
verbose warning for the condition for which
the validation fails. Useful for testing.}
}
\value{
a logical vector of length one indicating
               whether the input is a valid tokens object
}
\description{
A valid tokens object is a data frame or an object that
least two columns. The first column is called doc_id
and is a character vector with UTF-8 encoding. Document
ids must be unique. The second column is called token
and must also be a character vector in UTF-8 encoding.
Each individual token is represented by a single row in
the data frame. Addition token-level metadata columns
are allowed but not required.
}
\details{
The tests are run sequentially and the function returns,
with a warning if the warn flag is set, on the first test
that fails. We use this implementation because some tests
may fail entirely or be meaningless if the prior ones are
note passed. For example, if the tokens object does not
have a variable named "doc_id" it does not make sense to
check whether this column is a character vector.
}
\examples{
tokens <- data.frame(doc_id = c("doc1", "doc1", "doc1", "doc1",
                                "doc2",  "doc2", "doc2", "doc2",
                                "doc2", "doc2", "doc3", "doc3",
                                "doc3", "doc3", "doc3", "doc3"),
                     token = c("aujourd'hui", "maman", "est",
                               "morte", "it", "was", "a", "pleasure",
                               "to", "burn", "all", "this", "happened",
                               "more", "or", "less"))

tif_tokens_validate(tokens)

tokens$pos <- "NOUN"
tokens$NER <- ""
tokens$sentiment <- runif(16L)
tif_tokens_validate(tokens)
}
