% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg.R
\docType{package}
\name{tif-package}
\alias{tif}
\alias{tif-package}
\title{tif: Text Interchange Formats}
\description{
This package describes and validates formats for storing
common object arising in text analysis as native R objects.
Representations of a text corpus, document term matrix, and
tokenized text are included. The corpus and tokens objects
have multiple valid formats. Packages compliant with the
tif proposal should accept all valid formats and should
directly return, or provide conversion functions, for
converting outputs into at least one of the formats (when
applicable). The tokenized text format is extensible to
include other annotations such as part of speech tags and
named entities.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ropensci/tif}
  \item Report bugs at \url{http://github.com/ropensci/tif/issues}
}

}
